+++ Tests/Tests.Actors.BugFinding/Integration/ChordTests.cs
@@ -830,7 +830,7 @@ namespace Microsoft.Coyote.Actors.BugFinding.Tests
 
         [Theory(Timeout = 10000)]
         [InlineData(20)]
-        public void TestLivenessBugInChordProtocol(uint seed)
+        public void TestLivenessBugInChordProtocol(ulong seed)
         {
             var configuration = this.GetConfiguration();
             configuration.MaxUnfairSchedulingSteps = 200;
